sourcing common file

+++ components/common.sh
@@ -9,5 +9,14 @@ then
 fi
 }
 
+CHECK_STAT() {
+if [ $1 -ne 0 ]; then
+  echo -e "\e[31mFailed\e[0m"
+  exit 2
+  else
+  echo -e "\e[32m Success\e[0m"
+  fi
+
+}
 LOG=/tmp/roboshop.log
 rm -rf $LOG
